requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.0.5
django-auth-ldap==1.7.0
django-celery-beat==1.1.1
django-cors-headers==3.0.1
django-filter==1.1.0
django-mysql==2.4.1
django-redis==4.10.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ecdsa==0.13.2
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
paramiko==2.4.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyotp==2.2.7
python-ldap==3.1.0
pytz==2019.1
qrcode==6.1
redis==3.2.1
requests==2.21.0
six==1.12.0
sshpubkeys==3.1.0
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0